tests/wpt/tests/webmessaging/without-ports/026.html

<!doctype html>
<title>Cloning objects with getter properties</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
   var obj = {};
   var err = new Error("getter_should_propagate_exceptions");
   obj.__defineGetter__( "field", function(){ throw err });

   assert_throws_exactly(err, function() {
     postMessage(obj, '*');
  });
  this.done();
});
</script>
